refactor: navbar Icon과 text로 파일 분류
